requirements.txt
scrapy-zyte-api
scrapy-feedexporter-sftp